feat: admin -> admin-legacy | admin2 -> admin
Migrate admin to admin-legacy command
Migrate admin2 to admin command

The new admin is a refactoring to the CLI Metadata administration
interface.
The new admin uses the new securesystemslib Key format that is also
supported by RSTUF
