+++ repository_service_tuf/cli/admin/__init__.py
@@ -1,10 +1,25 @@
-# SPDX-FileCopyrightText: 2022-2023 VMware Inc
+# SPDX-FileCopyrightText: 2023-2024 Repository Service for TUF Contributors
 #
 # SPDX-License-Identifier: MIT
 
-from repository_service_tuf.cli import rstuf
+
+"""Alternative admin cli
+
+Provides alternative ceremony, metadata update, and sign admin cli commands.
+
+"""
+
+from repository_service_tuf.cli import click, rstuf
 
 
 @rstuf.group()  # type: ignore
 def admin():
     """Administrative Commands"""
+
+
+@admin.group()
+@click.pass_context
+def metadata(context):
+    """
+    Metadata management.
+    """
