+++ repository_service_tuf/cli/admin/update.py
@@ -1,8 +1,14 @@
+# SPDX-FileCopyrightText: 2023-2024 Repository Service for TUF Contributors
+#
+# SPDX-License-Identifier: MIT
+
 import json
+from copy import deepcopy
 from dataclasses import asdict
 
 import click
 from rich.markdown import Markdown
+from rich.prompt import Confirm
 from tuf.api.metadata import Metadata, Root
 
 # TODO: Should we use the global rstuf console exclusively? We do use it for
@@ -12,65 +18,70 @@ from tuf.api.metadata import Metadata, Root
 # https://rich.readthedocs.io/en/stable/console.html#console-api
 # https://rich.readthedocs.io/en/stable/console.html#capturing-output
 from repository_service_tuf.cli import console
-from repository_service_tuf.cli.admin2 import admin2
-from repository_service_tuf.cli.admin2.helpers import (
-    BinsRole,
-    CeremonyPayload,
+from repository_service_tuf.cli.admin import metadata
+from repository_service_tuf.cli.admin.helpers import (
+    EXPIRY_FORMAT,
     Metadatas,
-    Role,
-    Roles,
-    Settings,
+    UpdatePayload,
     _add_root_signatures_prompt,
     _configure_online_key_prompt,
     _configure_root_keys_prompt,
     _expiry_prompt,
-    _online_settings_prompt,
     _print_root,
     _root_threshold_prompt,
     _warn_no_save,
 )
 
 
-@admin2.command()  # type: ignore
+@metadata.command()  # type: ignore
+@click.argument("root_in", type=click.File("rb"))
 @click.option(
     "--save",
     "-s",
     is_flag=False,
-    flag_value="ceremony-payload.json",
-    help="Write json result to FILENAME (default: 'ceremony-payload.json')",
+    flag_value="update-payload.json",
+    help="Write json result to FILENAME (default: 'update-payload.json')",
     type=click.File("w"),
 )
-def ceremony(save) -> None:
-    """Bootstrap Ceremony to create initial root metadata and RSTUF config."""
-    console.print("\n", Markdown("# Metadata Bootstrap Tool"))
+def update(root_in, save) -> None:
+    """Update root metadata and bump version."""
+    console.print("\n", Markdown("# Metadata Update Tool"))
 
     if not save:
         _warn_no_save()
 
-    root = Root()
+    ###########################################################################
+    # Load root
+    # TODO: load from API
+    prev_root_md = Metadata[Root].from_bytes(root_in.read())
+    root = deepcopy(prev_root_md.signed)
 
     ###########################################################################
-    # Configure online role settings
-    console.print(Markdown("##  Online role settings"))
-    online = _online_settings_prompt()
+    # Configure root expiration
+    console.print(Markdown("## Root Expiration"))
 
-    console.print(Markdown("##  Root expiry"))
-    root_days, root_date = _expiry_prompt("root")
-    root.expires = root_date
-
-    roles = Roles(
-        Role(root_days),
-        Role(online.timestamp_expiry),
-        Role(online.snapshot_expiry),
-        Role(online.targets_expiry),
-        BinsRole(online.bins_expiry, online.bins_number),
+    expired = root.is_expired()
+    expiry_str = (
+        f"Root expire{'d' if expired else 's'} "
+        f"on {root.expires:{EXPIRY_FORMAT}}."
     )
+    if expired or Confirm.ask(
+        f"{expiry_str} Do you want to change the expiry date?", default=True
+    ):
+        _, date = _expiry_prompt("root")
+        root.expires = date
 
     ###########################################################################
-    # Configure Root Keys
+    # Configure root keys
     console.print(Markdown("## Root Keys"))
     root_role = root.get_delegated_role(Root.type)
-    root_role.threshold = _root_threshold_prompt()
+
+    threshold_str = f"Root signature threshold is {root_role.threshold}."
+    if Confirm.ask(
+        f"{threshold_str} Do you want to change the threshold?", default=False
+    ):
+        root_role.threshold = _root_threshold_prompt()
+
     _configure_root_keys_prompt(root)
 
     ###########################################################################
@@ -79,23 +90,26 @@ def ceremony(save) -> None:
     _configure_online_key_prompt(root)
 
     ###########################################################################
+    # Bump version
+    # TODO: check if metadata changed, or else abort? start over?
+    root.version += 1
+
+    ###########################################################################
     # Review Metadata
     console.print(Markdown("## Review"))
+    root_md = Metadata(root)
     _print_root(root)
     # TODO: ask to continue? or abort? or start over?
 
     ###########################################################################
     # Sign Metadata
     console.print(Markdown("## Sign"))
-    root_md = Metadata(root)
-    _add_root_signatures_prompt(root_md, None)
+    _add_root_signatures_prompt(root_md, prev_root_md.signed)
 
     ###########################################################################
     # Dump payload
     # TODO: post to API
+    payload = UpdatePayload(Metadatas(root_md.to_dict()))
     if save:
-        metadatas = Metadatas(root_md.to_dict())
-        settings = Settings(roles)
-        payload = CeremonyPayload(settings, metadatas)
         json.dump(asdict(payload), save, indent=2)
         console.print(f"Saved result to '{save.name}'")
