--- tests/unit/cli/admin/__init__.py
@@ -1,3 +1,0 @@
-# SPDX-FileCopyrightText: 2022-2023 VMware Inc
-#
-# SPDX-License-Identifier: MIT
