+++ tests/unit/cli/admin_legacy/test_import_artifacts.py
@@ -1,10 +1,15 @@
+# SPDX-FileCopyrightText: 2023-2024 Repository Service for TUF Contributors
+# SPDX-FileCopyrightText: 2022-2023 VMware Inc
+#
+# SPDX-License-Identifier: MIT
+
 import datetime
 from datetime import timezone
 
 import pretend
 import pytest
 
-from repository_service_tuf.cli.admin import import_artifacts
+from repository_service_tuf.cli.admin_legacy import import_artifacts
 
 
 class TestImportArtifactsFunctions:
@@ -78,7 +83,7 @@ class TestImportArtifactsFunctions:
             now=pretend.call_recorder(lambda a: fake_time)
         )
         monkeypatch.setattr(
-            "repository_service_tuf.cli.admin.import_artifacts.datetime",
+            "repository_service_tuf.cli.admin_legacy.import_artifacts.datetime",  # noqa
             fake_datetime,
         )
         succinct_roles = pretend.stub(
